pyproject.toml
[project]
name = "LangchainLearning"
version = "0.1.0"
description = "AI Engineer's Guide to LangChain. Full course on LangChain with OpenAI."
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-core>=0.3.15",
    "langchain-openai>=0.2.5",
    "langsmith>=0.1.139",
    "langchain-community>=0.3.5",
    "langchain>=0.3.7",
    "docarray>=0.40.0",
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "google-search-results>=2.4.2",
    "langchain-google-genai>=2.0.9",
    "matplotlib>=3.10.7",
    "huggingface-hub>=0.36.0",
    "langchain-huggingface>=1.0.0",
]
